+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ### [0.2.19] - 2024-06-04
 
 - better documente for `fav_core`
+- fix bugs in batch ops
+- fix ctrl-c not handled: handle `SIGINT` in `fav_core`'s Ext methods
+- auth: auto fetch after login
+- (un)track: show hint after (un)track 
+- pull: id not found hint improve
 
 ## [0.2.18] - 2024-06-04